+++ example/lib/main.dart
@@ -153,8 +153,8 @@ class SchemaOne extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return SchemaWidget(
-      config: SchemaConfig(
+    return Schema(
+      config: SchemaConfiguration(
         showBlocks: showBlockAreas.value,
         showGrid: showGridCells.value,
         initiateAxesScale: (constraints) => AxesScale(