+++ lib/src/block/block.dart
@@ -1,4 +1,4 @@
-//GENERATED BARREL FILE 
-export 'alignment.dart'; 
-export 'entity.dart'; 
-export 'tooling.dart'; 
+//GENERATED BARREL FILE
+export 'alignment.dart';
+export 'entity.dart';
+export 'tooling.dart';